# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Regression Hunter AI API
  version: 0.1.0
  description: |
    Predict, explain, and prevent software regressions before a pull request is merged.
    All APIs are versioned at `/v1`, authenticated with OIDC bearer tokens,
    and enforce repository/team entitlements.
  license:
    name: MIT
servers:
  - url: http://localhost:8001
    description: Local development
  - url: https://regression-hunter.example.com
    description: Production
security:
  - bearerAuth: []
paths:
  /v1/webhooks/github:
    post:
      summary: Receive GitHub App webhook events
      description: Handles pull_request, pull_request_review, and push events
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: GitHub webhook payload
      parameters:
        - name: X-Hub-Signature-256
          in: header
          required: true
          schema: { type: string }
        - name: X-GitHub-Event
          in: header
          required: true
          schema: { type: string }
        - name: X-GitHub-Delivery
          in: header
          required: true
          schema: { type: string }
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string }
                  run_id: { type: string, format: uuid }
                  run_key: { type: string }
        '403':
          description: Invalid signature
        '422':
          description: Unprocessable entity
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...